Rename the `test_client` submodule to `client` so that PyTest doesn't think it's an actual test
